week_3_data_warehouse/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
datetime
